requirements.txt
soco>=0.30.12
pylast>=6.0.0
python-dotenv>=1.2.1
